image-detection-service/requirements.txt
ultralytics
torch
torchvision
opencv-python-headless
fastapi
uvicorn
python-multipart
requests
Pillow
